+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/CommonAbstracts/types/ByInlineMessageId.kt
@@ -1,7 +1,0 @@
-package com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.types
-
-@Deprecated(
-    "Deprecated for the reason of creating of more obvious type interface",
-    ReplaceWith("InlineMessageAction", "com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.types.InlineMessageAction")
-)
-typealias ByInlineMessageId = InlineMessageAction