+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/bot/BaseRequestsExecutor.kt
@@ -4,14 +4,4 @@ import com.github.insanusmokrassar.TelegramBotAPI.utils.TelegramAPIUrlsKeeper
 
 abstract class BaseRequestsExecutor(
     protected val telegramAPIUrlsKeeper: TelegramAPIUrlsKeeper
-) : RequestsExecutor {
-    @Deprecated("Deprecated due to new TelegramAPIUrlKeeper API", ReplaceWith("telegramAPIUrlsKeeper.commonAPIUrl"))
-    protected val baseUrl: String
-        get() = telegramAPIUrlsKeeper.commonAPIUrl
-
-    @Deprecated("Deprecated due to new TelegramAPIUrlKeeper API")
-    constructor(
-        token: String,
-        hostUrl: String = "https://api.telegram.org"
-    ) : this (TelegramAPIUrlsKeeper(token, hostUrl))
-}
+) : RequestsExecutor